Cambio Toast por SweetAlert, junto con otras mejoras

+++ pages/biblioteca.vue
@@ -5,86 +5,80 @@
     <div class="search-section">
       <div class="search-form">
 
-        <input class="search-input" v-model="searchTerm" placeholder="Buscar por título o autor"
-          @input="filtrarLibros" />
-        <button @click="handleSubmit" class="boton-generico">Buscar</button>
+        <input class="search-input" v-model="searchTerm" placeholder="Buscar por título o autor" />
+        <button @click="buscarTexto" class="boton-generico">Buscar</button>
       </div>
 
-      <select class="boton-generico" v-model="ordenCalificacion" @change="filtrarLibros">
+      <select class="boton-generico" v-model="filter.ordenCalificacion">
         <option value="">Ordenar por calificación</option>
         <option value="asc">Ascendente</option>
         <option value="desc">Descendente</option>
       </select>
 
       <label>
-        <input type="checkbox" v-model="excluirSinReview" @change="filtrarLibros" />
+        <input type="checkbox" v-model="filter.excluirSinReview" />
         Excluir libros sin review
       </label>
     </div>
 
     <!-- Lista de libros -->
-    <div class="results-section">
-      <div v-for="libro in librosFiltrados" :key="libro.title" class="libro-card">
-        <img v-if="libro.coverImage" :title="libro.coverImage" :src="libro.coverImage" :alt="libro.title" />
 
-        <div class="books-grid">
-          <h3>{{ libro.title }}</h3>
-          <p><strong>Autor:</strong> {{ libro.author }}</p>
-          <p v-if="libro.publishYear"><strong>Año:</strong> {{ libro.publishYear }}</p>
-          <p v-if="libro.isbn"><strong>ISBN:</strong> {{ libro.isbn }}</p>
+    <div v-if="loading" class="results-section">
+      <span class="spinner-border spinner-border-sm" role="status"></span>
+      Guardando...
+    </div>
+    <div v-else v-for="libro in librosFiltrados" :key="libro.title" class="results-section">
+      <img v-if="libro.coverImage" :title="libro.coverImage" :src="libro.coverImage" :alt="libro.title" />
 
-          <div v-if="libro.review">
-            <strong>Review:</strong>
-            <p>{{ libro.review }}</p>
-          </div>
+      <div class="books-grid">
+        <h3>{{ libro.title }}</h3>
+        <p><strong>Autor:</strong> {{ libro.author }}</p>
+        <p v-if="libro.publishYear"><strong>Año:</strong> {{ libro.publishYear }}</p>
+        <p v-if="libro.isbn"><strong>ISBN:</strong> {{ libro.isbn }}</p>
 
-          <div v-if="libro.rating">
-            <strong>Calificación:</strong>
-            <span>{{ libro.rating }}/5</span>
-          </div>
+        <div v-if="libro.review">
+          <strong>Review:</strong>
+          <p>{{ libro.review }}</p>
         </div>
 
-        <!-- Botones de acción -->
-        <div class="acciones">
-          <button @click="editarLibro(libro)">Editar</button>
-          <button @click="eliminarLibro(libro)">Eliminar</button>
+        <div v-if="libro.rating">
+          <strong>Calificación:</strong>
+          <span>{{ libro.rating }}/5</span>
         </div>
       </div>
 
-      <!-- Mensaje si no hay libros -->
-      <div v-if="librosFiltrados.length === 0" class="sin-libros">
-        No se encontraron libros en tu biblioteca
+      <!-- Botones de acción -->
+      <div class="acciones">
+        <button type="button" class="btn btn-secondary btn-sm" @click="editarLibro(libro)">Editar</button>
+        <button type="button" class="btn btn-danger btn-sm" @click="eliminarLibro(libro)">Eliminar</button>
+
       </div>
     </div>
+
+    <!-- Mensaje si no hay libros -->
+    <div v-if="librosFiltrados.length === 0" class="sin-libros">
+      No se encontraron libros en tu biblioteca
+    </div>
+
 
     <!-- Modal de edición -->
     <div v-if="libroEditando" class="modal-edicion">
-    <div class="modal-contenido">
-      <h3>Editar {{ libroEditando.title }}</h3>
+      <div class="modal-contenido">
+        <h3>Editar {{ libroEditando.title }}</h3>
 
-      <textarea 
-        v-model="reviewEditado" 
-        placeholder="Escribe tu review" 
-        class="textarea-edicion"
-        maxlength="500" rows="6"
-      />
+        <textarea v-model="reviewEditado" placeholder="Escribe tu review" class="textarea-edicion" maxlength="500"
+          rows="6" />
 
-      <label>puntaje</label>
-      <input 
-        type="number" 
-        v-model="ratingEditado" 
-        min="1" 
-        max="5" 
-        placeholder="Calificación (1-5)" 
-        class="input-rating"
-      />
+        <label>puntaje</label>
+        <input type="number" v-model="ratingEditado" min="1" max="5" placeholder="Calificación (1-5)"
+          class="input-rating" />
 
-      <div class="modal-acciones">
-        <button @click="guardarEdicion" class="btn-guardar">Guardar</button>
-        <button @click="cancelarEdicion" class="btn-cancelar">Cancelar</button>
+        <div class="modal-acciones">
+          <button @click="guardarEdicion" class="btn-guardar">Guardar</button>
+          <button @click="cancelarEdicion" class="btn-cancelar">Cancelar</button>
+        </div>
       </div>
     </div>
-  </div>
   </div>
 </template>
 
@@ -96,13 +90,20 @@ definePageMeta({
 const libros = ref([])
 const librosFiltrados = ref([])
 const searchTerm = ref('')
-const ordenCalificacion = ref('')
-const excluirSinReview = ref(false)
+const loading = ref(false)
 const libroEditando = ref(null)
 const reviewEditado = ref('')
 const ratingEditado = ref(0)
 const toast = useToast()
+const filter = ref({
+  search: '',
+  ordenCalificacion: '',
+  excluirSinReview: false
+})
+
+//Plugins
 const { $api } = useNuxtApp()
+const { $swal } = useNuxtApp();
 
 // Cargar libros al iniciar
 onMounted(async () => {
@@ -111,41 +112,52 @@ onMounted(async () => {
 
 const cargarLibros = async () => {
   try {
+    loading.value = true
+
     const data = await $api(`/api/books/my-library`, {
       method: "GET",
     })
 
     librosFiltrados.value = data
+
   } catch (error) {
     toast.error({ title: 'Error!', message: error })
+  } finally {
+    setTimeout(() => {
+      loading.value = false
+    }, 300);
   }
 }
 
-const filtrarLibros = () => {
-  let filtered = [...libros.value]
+const filtrarLibros = async () => {
+  try {
 
-  // Filtrar por búsqueda
-  if (searchTerm.value) {
-    const term = searchTerm.value.toLowerCase()
-    filtered = filtered.filter(libro =>
-      libro.title.toLowerCase().includes(term) ||
-      libro.author.toLowerCase().includes(term)
-    )
+    loading.value = true
+
+    const params = {
+      excluirSinReview: filter.value.excluirSinReview,
+      ordenCalificacion: filter.value.ordenCalificacion,
+      search: filter.value.search
+    }
+
+    const cleanFilters = Object.fromEntries(
+      Object.entries(params).filter(([_, v]) => v !== undefined && v !== '')
+    );
+
+    const data = await $api(`/api/books/my-library`, {
+      method: "GET",
+      params: cleanFilters
+    })
+
+    librosFiltrados.value = data
+
+  } catch (error) {
+    toast.error({ title: 'Error!', message: 'Error al filtrar' })
+  } finally {
+    setTimeout(() => {
+      loading.value = false
+    }, 300);
   }
-
-  // Excluir sin review
-  if (excluirSinReview.value) {
-    filtered = filtered.filter(libro => libro.review && libro.review.trim() !== '')
-  }
-
-  // Ordenar por calificación
-  if (ordenCalificacion.value === 'asc') {
-    filtered.sort((a, b) => (a.rating || 0) - (b.rating || 0))
-  } else if (ordenCalificacion.value === 'desc') {
-    filtered.sort((a, b) => (b.rating || 0) - (a.rating || 0))
-  }
-
-  librosFiltrados.value = filtered
 }
 
 const editarLibro = (libro) => {
@@ -156,7 +168,7 @@ const editarLibro = (libro) => {
 
 const guardarEdicion = async () => {
   try {
-    if (reviewEditado.value === '' || ratingEditado.value === 0){
+    if (reviewEditado.value === '' || ratingEditado.value === 0) {
       return toast.warning({ title: 'Warning!', message: 'Complete los campos' })
     }
 
@@ -171,7 +183,7 @@ const guardarEdicion = async () => {
     // Actualizar localmente
     await cargarLibros();
     cancelarEdicion();
-    
+
   } catch (error) {
     console.error('Error editando libro:', error)
   }
@@ -184,8 +196,19 @@ const cancelarEdicion = () => {
 }
 
 const eliminarLibro = async (libro) => {
-
   try {
+
+    const result = $swal.fire({
+      title: "¿Estás seguro de eliminar este libro?",
+      text: "",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Sí, eliminar",
+      cancelButtonText: "Cancelar",
+    });
+
+    if (!(await result).isConfirmed) return;
+
     await $api(`/api/books/my-library/${libro._id}`, {
       method: 'DELETE'
     })
@@ -193,20 +216,26 @@ const eliminarLibro = async (libro) => {
     // Eliminar localmente
     await cargarLibros();
     cancelarEdicion();
+
   } catch (error) {
     console.error('Error eliminando libro:', error)
   }
 }
 
-const irAInicio = () => {
-  // Navegar al inicio
-  // router.push('/')
+const buscarTexto = () => {
+  if (searchTerm.value == '') {
+    toast.warning({ title: 'Warning!', message: 'Digite el nombre del libro' })
+  }
+  filter.value.search = searchTerm.value
 }
 
-const salir = () => {
-  // Lógica de logout
-  // auth.logout()
-}
+watch(
+  filter.value,
+  () => {
+    filtrarLibros()
+  }
+)
+
 </script>
 <style>
 .modal-edicion {
@@ -235,5 +264,9 @@ const salir = () => {
   margin: 0 0 1.5rem 0;
   color: #333;
   font-size: 1.5rem;
+}
+
+.acciones {
+  padding: 0.5em;
 }
 </style>
